feat: restructure and extract common class and func
